default to cleanup=False, fix bugs in maestro and orchset download methods

+++ mirdata/datasets/maestro.py
@@ -271,7 +271,7 @@ class Dataset(core.Dataset):
     def load_notes(self, *args, **kwargs):
         return load_notes(*args, **kwargs)
 
-    def download(self, partial_download=None, force_overwrite=False, cleanup=True):
+    def download(self, partial_download=None, force_overwrite=False, cleanup=False):
         """Download the dataset
 
         Args:
@@ -280,7 +280,6 @@ class Dataset(core.Dataset):
                 If None, all data is downloaded
             force_overwrite (bool):
                 If True, existing files are overwritten by the downloaded files. 
-                By default False.
             cleanup (bool):
                 Whether to delete any zip/tar files after extracting.
 
@@ -306,9 +305,24 @@ class Dataset(core.Dataset):
         # files get downloaded to a folder called maestro-v2.0.0
         # move everything up a level
         maestro_dir = os.path.join(self.data_home, "maestro-v2.0.0")
+        if not os.path.exists(maestro_dir):
+            logging.info(
+                "Maestro data not downloaded, because it probably already exists on your computer. "
+                + "Run .validate() to check, or rerun with force_overwrite=True to delete any "
+                + "existing files and download from scratch"
+            )
+            return
         maestro_files = glob.glob(os.path.join(maestro_dir, "*"))
 
         for fpath in maestro_files:
+            target_path = os.path.join(self.data_home, os.path.basename(fpath))
+            if os.path.exists(target_path):
+                logging.info(
+                    "{} already exists. Run with force_overwrite=True to download from scratch".format(
+                        target_path
+                    )
+                )
+                continue
             shutil.move(fpath, self.data_home)
 
         if os.path.exists(maestro_dir):
